factors out instrument tests
