+++ src/instrument/loudness.rs
@@ -14,39 +14,3 @@ pub fn loudness_normalization(frequency: f32) -> f32 {
     };
     normalization
 }
-
-#[cfg(test)]
-pub mod tests {
-    use super::*;
-
-    #[test]
-    fn test_loudness_normalization() {
-        let expected = loudness_normalization(0.0);
-        let result = 0.0;
-        assert_eq!(expected, result);
-
-        let expected = loudness_normalization(10.0);
-        let result = 0.0;
-        assert_eq!(expected, result);
-
-        let expected = loudness_normalization(100.0);
-        let result = 1.0;
-        assert_eq!(expected, result);
-
-        let expected = loudness_normalization(250.0);
-        let result = 0.5759918;
-        assert_eq!(expected, result);
-
-        let expected = loudness_normalization(500.0);
-        let result = 0.3794706;
-        assert_eq!(expected, result);
-
-        let expected = loudness_normalization(1000.0);
-        let result = 0.25;
-        assert_eq!(expected, result);
-
-        let expected = loudness_normalization(1500.0);
-        let result = 0.19584954;
-        assert_eq!(expected, result);
-    }
-}
