+++ src/instrument/mod.rs
@@ -1,3 +1,5 @@
-mod loudness;
-pub mod voice;
 pub mod oscillator;
+pub mod voice;
+mod loudness;
+#[cfg(test)]
+mod test;
